Move some network_service methods to sync_service
